+++ CHANGES.rst
@@ -20,6 +20,11 @@ API changes
  - ``nddata.utils`` is now a submodule. The affected utilities (``Cutout2D``,
    ...) must now be imported from there. PR #1
 
+ - ``NDData`` implements the ``__astropy_nddata__`` interface during instance
+   creation. PR #9
+
+ - ``NDData`` does not accept non-numerical arrays as data argument anymore. PR #9
+
 
 Bug fixes
 ^^^^^^^^^